+++ src/Humanizer.Tests/DateHumanizeTests.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Humanizer.Localisation;
+using Humanizer.Localization;
 using Xunit;
 
 namespace Humanizer.Tests
