+++ src/Humanizer.Tests/Localization/ar/NumberToWordsTests.cs
@@ -1,7 +1,7 @@
 ﻿using Xunit;
 using Xunit.Extensions;
 
-namespace Humanizer.Tests.Localisation.ar
+namespace Humanizer.Tests.Localization.ar
 {
     public class NumberToWordsTests : AmbientCulture
     {
