+++ src/Humanizer/Configuration/Configurator.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
-using Humanizer.Localisation;
+using Humanizer.Localization;
 
 namespace Humanizer.Configuration
 {
