+++ src/Humanizer/Localization/ResourceKeys.TimeSpanHumanize.cs
@@ -1,4 +1,4 @@
-﻿namespace Humanizer.Localisation
+﻿namespace Humanizer.Localization
 {
     public partial class ResourceKeys
     {
