+++ src/Humanizer/Localization/RussianFormatter.cs
@@ -1,4 +1,4 @@
-﻿namespace Humanizer.Localisation
+﻿namespace Humanizer.Localization
 {
     internal class RussianFormatter : DefaultFormatter
     {
